avoid skipping validation checks for numberOfShards setting when building IndexMetadata
With split changes, we've added SplitShardsMetadata. This object is instantiated in IndexMetadata.Builder' flow using the numberOfShards. We were invoking numberOfShards() which skips the validation check for the INDEX_NUMBER_OF_SHARDS_SETTING setting as it directly fetches the setting via `settings.getAsInt()`.

Changing this to fetch this setting via INDEX_NUMBER_OF_SHARDS_SETTING.get(settings) so that validations work correctly.

Signed-off-by: Paras Jain <parasjain.jain@outlook.com>